+++ xarray/util/print_versions.py
@@ -107,7 +107,6 @@ def show_versions(file=sys.stdout):
         ("zarr", lambda mod: mod.__version__),
         ("cftime", lambda mod: mod.__version__),
         ("nc_time_axis", lambda mod: mod.__version__),
-        ("PseudoNetCDF", lambda mod: mod.__version__),
         ("iris", lambda mod: mod.__version__),
         ("bottleneck", lambda mod: mod.__version__),
         ("dask", lambda mod: mod.__version__),